+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/ContentManagerPlugin.java
@@ -52,12 +52,14 @@ import java.util.function.Supplier;
 import com.wazuh.contentmanager.cti.catalog.index.ConsumersIndex;
 import com.wazuh.contentmanager.cti.catalog.service.SpaceService;
 import com.wazuh.contentmanager.cti.console.CtiConsole;
+import com.wazuh.contentmanager.engine.service.EngineService;
 import com.wazuh.contentmanager.engine.service.EngineServiceImpl;
 import com.wazuh.contentmanager.jobscheduler.ContentJobParameter;
 import com.wazuh.contentmanager.jobscheduler.ContentJobRunner;
 import com.wazuh.contentmanager.jobscheduler.jobs.CatalogSyncJob;
 import com.wazuh.contentmanager.rest.service.*;
 import com.wazuh.contentmanager.settings.PluginSettings;
+import com.wazuh.contentmanager.utils.MockEngineService;
 
 /** Main class of the Content Manager Plugin */
 public class ContentManagerPlugin extends Plugin
@@ -71,7 +73,7 @@ public class ContentManagerPlugin extends Plugin
     private CtiConsole ctiConsole;
     private Client client;
     private CatalogSyncJob catalogSyncJob;
-    private EngineServiceImpl engine;
+    private EngineService engine;
     private SpaceService spaceService;
 
     /**
@@ -122,7 +124,11 @@ public class ContentManagerPlugin extends Plugin
         runner.registerExecutor(CatalogSyncJob.JOB_TYPE, this.catalogSyncJob);
 
         // Initialize Engine service
-        this.engine = new EngineServiceImpl();
+        if (PluginSettings.getInstance().isEngineMockEnabled()) {
+            this.engine = new MockEngineService();
+        } else {
+            this.engine = new EngineServiceImpl();
+        }
 
         // Initialize Space Service
         this.spaceService = new SpaceService(this.client);
@@ -326,6 +332,7 @@ public class ContentManagerPlugin extends Plugin
                 PluginSettings.CONTENT_CONSUMER,
                 PluginSettings.IOC_CONTEXT,
                 PluginSettings.IOC_CONSUMER,
+                PluginSettings.ENGINE_MOCK_ENABLED,
                 PluginSettings.CREATE_DETECTORS);
     }